rev_b/sfw/scripts/requirements.txt
pyserial == 3.5
numpy == 1.20.1